--- src/js/modal-movie.js
@@ -1,7 +1,0 @@
-refs.modalOpen.addEventListener('click', onOpenModal);
-
-// function renderModal(data) {
-//   appendMovieMarcup(data, modalMovieTpl);
-// }
-
-// function appendMovieMarcup(data, modalMovieTpl) {}
