+++ src/js/refs.js
@@ -1,5 +1,11 @@
-const modalOpen = document.querySelector('.modal-open');
+const modalFilmOpen = document.querySelector('.modal-film-open');
+const modalTeamOpen = document.querySelector('.modal-team-open');
+const btnWatched = document.querySelector('.btn-watch');
+const btnQueue = document.querySelector('.btn-queue');
 
 export default {
-  modalOpen,
+  modalFilmOpen,
+  modalTeamOpen,
+  btnWatched,
+  btnQueue,
 };
